feat: Refactor DateRangeSelector to use constants from dateRangeConstants and improve test imports
